# api.yml
...
swagger: '2.0'
info:
  version: 1.0.0
  title: api-activos-fijo Microservice
  description: Java Microservice + MySQL Database
  contact:
    name: Jhon alexander chacon torres
    email: jhachacont@correo.udistrital.edu.co
  license:
    name: PruebaConocimiento
host: localhost
basePath: /api-activos-fijo-ms/resources
schemes:
- http
consumes:
- application/json
produces:
- application/json
paths:
  /tipo:
    get:
      summary: retorna todos los registros del tipo
      operationId: getAllTipo
      parameters:
        - name: firstResult
          in: query
          description: el primer resultado a retornar
          required: false
          type: integer
        - name: maxResult
          in: query
          description: el maximo numero de resultados a retornar
          required: false
          type: integer
      responses:
        200:
          description: Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/Tipo'
        406:
          description: Not Acceptable (Invalid Params)
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
    post:
      summary: crear nuevos tipos
      operationId: createTipo
      parameters:
      - name: Tipo
        in: body
        required: true
        description: Objeto Tipo a crear
        schema:
          $ref: '#/definitions/Tipo'
      responses:
        201:
          description: Respuesta para la creacion del objeto
          schema:
            $ref: '#/definitions/Tipo'
        400:
          description: Bad Request falta un dato
          schema:
            $ref: '#/definitions/ErrorModel'
        403:
          description: Forbidden el dato ya existe
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
    put:
      summary: Actualizar un objeto de tipo
      operationId: updateTipo
      parameters:
      - name: tipo
        in: body
        required: true
        description: Objeto tipo a actualizar
        schema:
          $ref: '#/definitions/Tipo'
      responses:
        200:
          description: Ok
          schema:
            $ref: '#/definitions/Tipo'
        400:
          description: Bad Request falta un dato
          schema:
            $ref: '#/definitions/ErrorModel'
        404:
          description: Not Found no existe el dato
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: unexpected error
          schema:
            $ref: '#/definitions/ErrorModel'
    delete:
      summary: Eliminar tipo segun llave o codigo
      operationId: deleteTipo
      parameters:
      - name: codigo
        in: query
        required: true
        description: codigo o lleve del tipo
        type: string
      responses:
        200:
          description: Ok
          schema:
            $ref: '#/definitions/Tipo'
        400:
          description: Bad Request falta un dato
          schema:
            $ref: '#/definitions/ErrorModel'
        404:
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: unexpected error
          schema:
            $ref: '#/definitions/ErrorModel'
  /tipo/byCodigo:
    get:
      summary: Retorna un tipo segun su codigo
      operationId: getTipoByCode
      parameters:
      - name: codigo
        in: query
        description: codigo del tipo
        required: true
        type: string
      responses:
        200:
          description: Ok
          schema:
            $ref: '#/definitions/Tipo'
        400:
          description: Bad Request falta un dato
          schema:
            $ref: '#/definitions/ErrorModel'
        404:
          description: Not Found no existe el dato
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: unexpected error
          schema:
            $ref: '#/definitions/ErrorModel'
...
  /personas:
    get:
      summary: retorna todos los registros de personas registrados
      operationId: getAllPersonas
      parameters:
      - name: firstResult
        in: query
        description: el primer resultado a retornar
        required: false
        type: integer
      - name: maxResult
        in: query
        description: el maximo numero de resultados a retornar
        required: false
        type: integer
      responses:
        200:
          description: Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/Personas'
        406:
          description: Not Acceptable (Invalid Params)
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
    post:
      summary: crear nuevas personas
      operationId: createPersonas
      parameters:
      - name: personas
        in: body
        required: true
        description: Objeto personas a crear
        schema:
          $ref: '#/definitions/Personas'
      responses:
        201:
          description: Respuesta para la creacion del objeto
          schema:
            $ref: '#/definitions/Personas'
        400:
          description: Bad Request falta un dato
          schema:
            $ref: '#/definitions/ErrorModel'
        403:
          description: Forbidden el dato ya existe
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
    put:
      summary: Actualizar un objeto de personas
      operationId: updatePersonas
      parameters:
      - name: personas
        in: body
        required: true
        description: Objeto personas a actualizar
        schema:
          $ref: '#/definitions/Personas'
      responses:
        200:
          description: Ok
          schema:
            $ref: '#/definitions/Personas'
        400:
          description: Bad Request falta un dato
          schema:
            $ref: '#/definitions/ErrorModel'
        404:
          description: Not Found no existe el dato
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: unexpected error
          schema:
            $ref: '#/definitions/ErrorModel'
    delete:
      summary: Eliminar personas segun llave o documento
      operationId: deletePersonas
      parameters:
      - name: documento
        in: query
        required: true
        description: documento o lleve del estado actual
        type: string
      responses:
        200:
          description: Ok
          schema:
            $ref: '#/definitions/Personas'
        400:
          description: Bad Request falta un dato
          schema:
            $ref: '#/definitions/ErrorModel'
        404:
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: unexpected error
          schema:
            $ref: '#/definitions/ErrorModel'
  /personas/byDocument:
    get:
      summary: Retorna una persona segun su documento
      operationId: getPersonasByDocumento
      parameters:
      - name: documento
        in: query
        description: documento de la persona
        required: true
        type: string
      responses:
        200:
          description: Ok
          schema:
            $ref: '#/definitions/Personas'
        400:
          description: Bad Request falta un dato
          schema:
            $ref: '#/definitions/ErrorModel'
        404:
          description: Not Found no existe el dato
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: unexpected error
          schema:
            $ref: '#/definitions/ErrorModel'
...
definitions:
  Tipo:
    type: object
    required: ["codigo", "descripcion"]
    properties:
      codigo:
        type: string
      descripcion:
        type: string
...
  Personas:
    type: object
    required: ["documento", "nombres", "apellidos", "telefono", "correo", "direccion"]
    properties:
      documento:
        type: string
      nombres:
        type: string
      apellidos:
        type: string
      telefono:
        type: string
      correo:
        type: string
      direccion:
        type: string
...
  ErrorModel:
    type: object
    required:
    - code
    - description
    - id
    properties:
      code:
        type: integer
      description:
        type: object
        properties:
          key:
            type: string
          value:
            type: string
      id:
        type: string